Move test code to test files
