+++ src/ratchet_tree.cpp
@@ -731,14 +731,4 @@ operator>>(tls::istream& in, RatchetTree& obj)
   return in;
 }
 
-namespace test {
-
-const RatchetTreeNodeVector&
-TestRatchetTree::nodes() const
-{
-  return _nodes;
-}
-
-} // namespace test
-
 } // namespace mls
