+++ src/session.cpp
@@ -200,52 +200,4 @@ Session::current_state()
   return _state.at(_current_epoch);
 }
 
-/////
-
-namespace test {
-
-uint32_t
-TestSession::index() const
-{
-  return current_state().index().val;
-}
-
-epoch_t
-TestSession::current_epoch() const
-{
-  return _current_epoch;
-}
-
-CipherSuite
-TestSession::cipher_suite() const
-{
-  return current_state().cipher_suite();
-}
-
-bytes
-TestSession::current_epoch_secret() const
-{
-  return current_state().epoch_secret();
-}
-
-bytes
-TestSession::current_application_secret() const
-{
-  return current_state().application_secret();
-}
-
-bytes
-TestSession::current_confirmation_key() const
-{
-  return current_state().confirmation_key();
-}
-
-bytes
-TestSession::current_init_secret() const
-{
-  return current_state().init_secret();
-}
-
-} // namespace test
-
 } // namespace mls
